Fix global module fragment warnings from g++ 15.2

+++ UnitTests/DiracIntegral.C
@@ -112,7 +112,7 @@ TEST_F(DiracIntegralTests, SlaterOverlap)
         SMatrix<double> S=oi->Overlap();
         {
             RVec d=S.GetDiagonal();
-            for (int i=1;i<=d.size()/2;i++) EXPECT_NEAR(d(i),1.0,1e-15);
+            for (size_t i=1;i<=d.size()/2;i++) EXPECT_NEAR(d(i),1.0,1e-15);
         }
         // cout << std::fixed << std::setprecision(3) << std::setw(6) << S << S1 << endl;
         const IrrepBasisSet<double>* l=GetLarge(oi);
@@ -134,7 +134,7 @@ TEST_F(DiracIntegralTests, GaussianOverlap)
         SMatrix<double> S=oi->Overlap();
         {
             RVec d=S.GetDiagonal();
-            for (int i=1;i<=d.size()/2;i++) EXPECT_NEAR(d(i),1.0,1e-15);
+            for (size_t i=1;i<=d.size()/2;i++) EXPECT_NEAR(d(i),1.0,1e-15);
         }
         // cout << std::fixed << std::setprecision(3) << std::setw(6) << S << S1 << endl;
         const IrrepBasisSet<double>* l=GetLarge(oi);
